+++ errors/definitions/multi_error.go
@@ -14,6 +14,12 @@ import (
 type MultiError struct {
 	errs []error
 
+	// currentOutput keeps tally of the total number of bytes in the error message.
+	currentOutput int
+
+	// OutputCap is the maximum number of bytes to be used for error messages. If 0, we don't cap.
+	OutputCap int
+
 	// Prefix will be printed before the errors in this MultiError.
 	Prefix string
 	// Separator will separate the errors in this MultiError.
@@ -23,11 +29,30 @@ type MultiError struct {
 
 // Add adds a new error to this MultiError.
 func (m *MultiError) Add(err error) {
-	if err == nil {
+	if err == nil || err == m {
 		return
 	}
 
-	m.errs = append(m.errs, err)
+	if m.shouldAddErr(err) {
+		m.errs = append(m.errs, err)
+	}
+}
+
+// shouldAddErr checks if we should add the error to the MultiError by checking if we have hit the output cap.
+func (m *MultiError) shouldAddErr(err error) bool {
+	// Should always add err when OutputCap is 0 - ie is considered not set and infinite.
+	if m.OutputCap == 0 {
+		return true
+	}
+
+	// We add the length always here even if we don't add so that we always block from when the first error goes over
+	// the limit. Doing this in alternative ways could result in errors missing but nothing to indicate that there are
+	// missing values; ie the output cap limit hit message in Error().
+	if m.currentOutput <= m.OutputCap {
+		m.currentOutput += len(err.Error())
+	}
+
+	return m.currentOutput <= m.OutputCap
 }
 
 func (m *MultiError) Error() string {
@@ -52,6 +77,10 @@ func (m *MultiError) Error() string {
 	}
 
 	errStr.WriteString(m.errs[len(m.errs)-1].Error())
+
+	if m.currentOutput > m.OutputCap {
+		errStr.WriteString("; error message output cap hit - not all errors are shown")
+	}
 
 	return errStr.String()
 }